# swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Welcome to the ORY Hydra HTTP API documentation. You will find documentation for all HTTP APIs here.",
    "title": "ORY Hydra",
    "version": "latest"
  },
  "basePath": "/",
  "paths": {
...
    "/identifier": {
      "get": {
        "description": "List all the data identifiers with the client_id.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "dataIdentifier"
        ],
        "summary": "List all the data identifiers",
        "operationId": "listDataIdentifiers",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ClientId",
            "name": "client_id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Tag",
            "name": "tag",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Metadata",
            "name": "metadata",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ListIdentifiersResp"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      },
      "post": {
        "description": "Create a new data identifier if you have the valid license(apiKey). The privateKey and publicKey are returned in response.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "dataIdentifier"
        ],
        "summary": "Create a data identifier",
        "operationId": "createDataIdentifier",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/DataIdentifierResp"
          },
          "400": {
            "$ref": "#/responses/jsonError"
          },
          "401": {
            "$ref": "#/responses/jsonError"
          },
          "403": {
            "$ref": "#/responses/jsonError"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      }
    },
    "/identifier/{id}": {
      "get": {
        "description": "Delete the data identifier with the identifierID.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "dataIdentifier"
        ],
        "summary": "Delete a data identifier",
        "operationId": "dataIdentifierID",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/jsonError"
          },
          "400": {
            "$ref": "#/responses/jsonError"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      },
      "delete": {
        "description": "Delete the data identifier with the identifierID.",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "dataIdentifier"
        ],
        "summary": "Delete a data identifier",
        "operationId": "dataIdentifierID",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/jsonError"
          },
          "400": {
            "$ref": "#/responses/jsonError"
          },
          "404": {
            "$ref": "#/responses/jsonError"
          },
          "500": {
            "$ref": "#/responses/jsonError"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Identifier": {
      "type": "object",
      "properties": {
        "authAddress": {
          "type": "string",
          "x-go-name": "AuthAddress"
        },
        "categoryID": {
          "type": "string",
          "x-go-name": "CategoryID"
        },
        "dataAddress": {
          "type": "string",
          "x-go-name": "DataAddress"
        },
        "dataDigest": {
          "type": "string",
          "x-go-name": "DataDigest"
        },
        "dataSignature": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint8"
          },
          "x-go-name": "DataSignature"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Metadata"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "subscriptionType": {
          "$ref": "#/definitions/SubscriptionType"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Tags"
        }
      },
      "x-go-package": "github.com/ory/hydra/identifier"
    },
...
    "SubscriptionType": {
      "type": "string",
      "x-go-package": "github.com/ory/hydra/subscription"
    },
...
  },
  "responses": {
    "DataIdentifierResp": {
      "description": "Data identifier information response are sent when the operation succeeds.",
      "schema": {
        "$ref": "#/definitions/Identifier"
      }
    },
...
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    },
    "oauth2": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://hydra.demo.ory.sh/oauth2/auth",
      "tokenUrl": "https://hydra.demo.ory.sh/oauth2/token",
      "scopes": {
        "offline": "A scope required when requesting refresh tokens (alias for `offline_access`)",
        "offline_access": "A scope required when requesting refresh tokens",
        "openid": "Request an OpenID Connect ID Token"
      }
    }
  },
...
}